requirements.txt
streamlit
pdfplumber
docx
pandas
openai==1.10.0
python-docx
python-dateutil
openpyxl